+++ internal/sdkprovider/service/influxdb/influxdb.go
@@ -14,8 +14,31 @@ func influxDBSchema() map[string]*schema.Schema {
 		Type:        schema.TypeList,
 		Computed:    true,
 		Description: "InfluxDB server provided values",
+		MaxItems:    1,
+		Optional:    true,
+		Sensitive:   true,
 		Elem: &schema.Resource{
 			Schema: map[string]*schema.Schema{
+				"uris": {
+					Type:        schema.TypeList,
+					Computed:    true,
+					Description: "InfluxDB server URIs.",
+					Optional:    true,
+					Elem: &schema.Schema{
+						Type: schema.TypeString,
+					},
+				},
+				"username": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "InfluxDB username",
+				},
+				"password": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "InfluxDB password",
+					Sensitive:   true,
+				},
 				"database_name": {
 					Type:        schema.TypeString,
 					Computed:    true,