+++ CHANGELOG.md
@@ -12,6 +12,7 @@ nav_order: 1
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
 - Fix `ip_filter`, `ip_filter_string`, and `ip_filter_object` crash when receive an unexpected type
+- Add missing connection info fields to various service types
 
 ## [4.17.0] - 2024-05-07